Add Month salary endpoint

+++ server-app/Timesheet/Timesheet.Core/Models/WorkedHours/AddWorkedHourModel.cs
@@ -7,6 +7,6 @@ namespace Timesheet.Core.Models.WorkedHours
         public int ProjectId { get; set; }
         public int ProjectTaskId { get; set; }
         public DateTime Date { get; set; }
-        public int HoursQuantity { get; set; }
+        public decimal HoursQuantity { get; set; }
     }
 }
